chore: 💳 Unify card code
