+++ src/components/designSystem/index.ts
@@ -1,7 +1,6 @@
 export * from './Alert'
 export * from './Avatar'
 export * from './Button'
-export * from './Card'
 export * from './Dialog'
 export * from './Icon'
 export * from './InfiniteScroll'
