+++ src/styles/designSystem/index.ts
@@ -1,4 +1,5 @@
 export * from './ButtonGroup'
+export * from './Card'
 export * from './List'
 export * from './MenuPopper'
 export * from './PageHeader'
